Prova_BD/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
pymongo>=4.5
motor>=3.1
dnspython>=2.3
redis>=4.2.0
pydantic
aioredis
python-dotenv
redis>=4.2.0